requirements.txt
Django==1.8.3
argparse==1.2.1
djangorestframework==3.2.0
wsgiref==0.1.2
